% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_as.R
\name{as_tag}
\alias{as_tag}
\alias{as_tag_adverb}
\title{Convert to a tag or a tag_adverb}
\usage{
as_tag(f)

as_tag_adverb(f)
}
\arguments{
\item{f}{a function operator (adverb)}
}
\description{
Convert without loss of information
}
