% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bathy}
\alias{bathy}
\title{Bathymetry example}
\format{
A matrix
}
\usage{
bathy
}
\description{
Example of bathymetry used for a Pacific cod in the Alrutian Islands.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{details for bathymetry
}\if{html}{\out{</div>}}
}
\keyword{datasets}
