% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_general.R
\name{CALL}
\alias{CALL}
\alias{T_ARGS}
\alias{F_ARGS}
\alias{T_FORMALS}
\alias{F_FORMALS}
\title{Pattern helpers}
\usage{
CALL(eval = TRUE, type = c("expanded", "unexpanded", "raw"))

T_ARGS(eval = TRUE)

F_ARGS(eval = TRUE, type = c("expanded", "unexpanded", "raw"))

T_FORMALS()

F_FORMALS()
}
\arguments{
\item{eval}{wether to evaluate the call or arguments}

\item{type}{type of call. \code{"expanded"} uses \code{match.call()}, \code{"unexpanded"}
uses \code{match.call(expand.dots=FALSE)}, \code{"raw"} uses \code{sys.call()}}
}
\description{
These functions should only be used inside of tag definitions.
}
\details{
\describe{
\item{\code{f}}{input function}
\item{\code{CALL()}}{return evaluated or unevaluated call}
\item{\code{F_ARGS()}}{return evaluated or unevaluated arguments of input function}
\item{\code{T_ARGS()}}{return evaluated or unevaluated arguments of input function}
\item{\code{F_FORMALS()}}{return the input function's formals}
\item{\code{T_FORMALS()}}{return he tag's formals}
\item{\code{NEW_FORMALS()}}{The formals of the \code{tag}, or additional formals of
the \code{tag_adverb}}
}
}
