% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod_ex}
\alias{cod_ex}
\title{Data to demonstrate Hidden Markov Model of a Pacific Cod}
\description{
Data sufficient to replicate some features of the application of a
Hidden Markov model to Aleutian Islands Pacific cod
}
\details{
\itemize{
\item \code{bathy} Stars object containing bathymetry
\item \code{L} Stars object containing daily likelihoods
}
}
\keyword{data}
