% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_kern.R
\name{conv_kern}
\alias{conv_kern}
\title{Create a convolution kernel}
\usage{
conv_kern(
  sigma = 1.4,
  k = c("gaussian", "LoG", "sharpen", "laplacian", "emboss", "sobel")
)
}
\arguments{
\item{sigma}{The numeric value of standard deviation for the Gaussian or LoG filter}

\item{k}{character value:
"gaussian" for Gaussian kernel,
"LoG" for Laplacian of Gaussian kernel,
"sharpen" for 3x3 convolution matrix for sharpening edges, or
"laplacian" for a 3x3 convolution matrix.}
}
\value{
A numeric convolution matrix
}
\description{
This function creates the convolution kernel for applying a filter to an array/matrix
\itemize{
\item spatialfil package is not available anymore, need to load function manually
}
}
\examples{
# Not run:
# creates a convolution kernel with Gaussian function and sigma = 1.4
K <- conv_kern(sigma = 1.4, k = 'gaussian')
raster::image(K$matrix) # Visualize the convolution kernel
}
