% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_dollar.R
\name{dollar}
\alias{dollar}
\alias{$.tag_adverb}
\alias{$.tag}
\title{tag and tag_adverb methods for $}
\usage{
\method{$}{tag_adverb}(e1, e2)

\method{$}{tag}(e1, e2)
}
\arguments{
\item{e1}{lhs}

\item{e2}{rhs}
}
\description{
\emph{tag} provides \code{tag} and \code{tag_adverb} methods for \code{$} which are essentially
syntactic sugar, and the additional advantage to provide autocomplete in some
cases.
}
\details{
\code{$.tag_adverb} makes it possible to write \code{some_tag(tag_arg)$fun(fun_arg)}
instead of \code{some_tag(tag_arg)(fun)(fun_arg)}.

\code{$.tag} makes it possible to write \code{some_tag$fun(fun_arg, tag_arg)}
instead of \code{some_tag()(fun)(fun_arg, tag_arg)}.

A additional advantage of \code{$.tag} is that autocomplete will be available to the
arguments of the manufactured function.
}
\examples{
\dontrun{
library(tags)
# These do the same, thanks to `$.tag_adverb`
trying(.else = "hi", .silent = TRUE)(paste)("hello", world)
trying(.else = "hi", .silent = TRUE)$paste("hello", world)

# These do the same, thanks to `$.tag`
trying()(paste)("hello", world, .else = "hi", .silent = TRUE)
trying$paste("hello", world, .else = "hi", .silent = TRUE)
}
}
