% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_general.R, R/07_p_fs.R
\name{dollar_methods}
\alias{dollar_methods}
\alias{$.tag}
\alias{$.fseq}
\alias{$.tagged_p}
\title{dollar methods for tag functions and tagged object}
\usage{
\method{$}{tag}(e1, e2)

\method{$}{fseq}(e1, e2)

\method{$}{tagged_p}(e1, e2)
}
\arguments{
\item{e1}{lhs}

\item{e2}{rhs}
}
\description{
tag functions can be used using the dollar notation so for \code{..foo$x}
is equivalent to \code{..foo(x)}. The resulting object may have their own
dollar method as well.
}
