% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_at.R
\name{..at}
\alias{..at}
\title{..at tag to be able to use \code{at()} inside of \code{summarize} and \code{mutate}}
\usage{
..at(f)
}
\arguments{
\item{f}{function}
}
\description{
..at tag to be able to use \code{at()} inside of \code{summarize} and \code{mutate}
}
\examples{
\dontrun{
library(dplyr)
..at$summarize(iris, at(vars(starts_with("Sepal")),lst(mean,median)))
mtcars \%>\%
  group_by(gear) \%>\%
  ..at$summarize(at(c("disp","hp"), funs(mean), .keys = vars(cyl)))
}
}
