% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_chr.R
\name{..chr}
\alias{..chr}
\title{..chr tag to use arithmetic operators on characters}
\usage{
..chr(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a tagged_chr object
}
\description{
Use \code{+} to append without separator (or to apply a function), \code{-} to prepend,
\code{*} to duplicate the string, and \code{/} to subset the elements that satisfy
the regex given as a rhs.
}
\examples{
x <- "hello"
..chr$x + "world"
..chr$x + "world" - "oh"
..chr$x + "world" - "oh" + toupper + unclass
..chr$x * 3
x <- c("hello","sup", "hey","mornin")
..chr$x / "^h"
}
