% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_dt.R
\name{..dt}
\alias{..dt}
\title{..dt tag to use data.table syntax without explicit conversion}
\usage{
..dt(data)
}
\arguments{
\item{data}{object coercible to data.table}
}
\value{
an object of same class as input
}
\description{
Will return an object of same class as input
}
\examples{
iris2 <- head(iris,2)
..dt$iris2[,Species2:=toupper(Species)]
}
