% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_fn.R
\name{..fn}
\alias{..fn}
\title{..fn tag to be able to use formula notation on FUN arguments}
\usage{
..fn(f)
}
\arguments{
\item{f}{function to tag}
}
\value{
a modified function
}
\description{
inspired by \code{gsubfn::fn}, \code{..fn} is similar but use \code{rlang::as_function}
whre \emph{gsubfn} uses its own API.
}
\examples{
\dontrun{
..fn(sapply)(iris, ~class(.))
..fn$sapply(iris, ~class(.))
sapply2 <- ..fn$sapply
sapply2(iris, ~class(.))
}
}
