% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_p_fs.R
\name{..fs}
\alias{..fs}
\title{..fs tag to start a dollar piped functional sequence}
\usage{
..fs(x)
}
\arguments{
\item{x}{a function}
}
\value{
a tagged function
}
\description{
Based on \code{\%>\%} from the package magrittr
}
\examples{
\dontrun{
..fs(head)$dim()(iris)
..fs$head(2)$gsub("h","X",.)(c("hello","hi","foo"))
}
}
