% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_grp.R
\name{..grp}
\alias{..grp}
\title{..grp tag to group and ungroup around another operation}
\usage{
..grp(fun)
}
\arguments{
\item{fun}{The function to tag}
}
\value{
a tagged function
}
\description{
The tagged function gains a parameter \code{.by}, groups that existed before
are preserved. One can group by a computed variable, it will be kept if
named, dropped otherwise.
}
\examples{
library(dplyr)
..grp(summarize)(iris,meanSL = mean(Sepal.Length), .by="Species")
..grp$summarize(iris,meanSL = mean(Sepal.Length), .by="Species")
..grp$summarize_all(iris,mean, .by="Species")
..grp$slice(iris,1, .by="Species")
..grp$summarize(iris,meanSW = mean(Sepal.Width),
  .by= vars(Species, Sepal.Width > 3.2))
..grp$summarize(iris,meanSW = mean(Sepal.Width), .by= vars(
  Species, long_sepal = Sepal.Width > 3.2))
}
