% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_incr.R
\name{..incr}
\alias{..incr}
\title{..incr tag to increment or decrement numeric variables}
\usage{
..incr(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector
}
\description{
Use unary of binary \code{-} and \code{+} to increment or decrement of \code{1} or amount
given on the left hand side.
}
\examples{
\dontrun{
x <- 3
+..incr$x
..incr$x + 5
-..incr$x
..incr$x - 2
}
}
