% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ip.R
\name{..ip}
\alias{..ip}
\title{..ip tag to modify objects in place}
\usage{
..ip(f)
}
\arguments{
\item{f}{a function that returns a modified object, fed as the first
parameter}
}
\description{
..ip tag to modify objects in place
}
\examples{
iris2 <- iris
..ip$head(iris2,2)
iris2
..ip$transform(iris2, Species2 = toupper(Species))
iris2
}
