% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_lbd.R
\name{..lbd}
\alias{..lbd}
\title{..lbd tag to apply function on argument when argument is a formula}
\usage{
..lbd(f)
}
\arguments{
\item{f}{function}
}
\description{
Useful in functions such as \code{transform}, \code{mutate}, \code{summarize}.
}
\examples{
..lbd$transform(head(iris,2), Petal.Width = ~1000*(.), Species = ~toupper(.))
}
