% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_p_fs.R
\name{..p}
\alias{..p}
\title{..p tag to start a dollar pipe sequence}
\usage{
..p(x)
}
\arguments{
\item{x}{an object}
}
\description{
..p tag to start a dollar pipe sequence
}
\examples{
\dontrun{
..p$iris$head$Filter(is.numeric,.)$dim()
}
}
