% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_try.R
\name{..try}
\alias{..try}
\title{..try tag to try a call or if failure call alternate call}
\usage{
..try(f)
}
\arguments{
\item{f}{function}
}
\description{
The expression stored in \code{else} will be executed if
}
\examples{
..try$paste("hello","world", .else = "hi")
..try$paste("hello", world, .else = "hi")
..try$paste("hello", world, .else = "hi", .silent = TRUE)
}
