% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_w.R
\name{..w}
\alias{..w}
\title{..w tag for compact subsetting and evaluation within objects}
\usage{
..w(x)
}
\arguments{
\item{x}{object}
}
\value{
object of class \code{tagged_w}
}
\description{
..w tag for compact subsetting and evaluation within objects
}
\examples{
\dontrun{
..w$iris[nrow(.),]
vec1 <- c(a=1,b=2,c=3,d=4)
..w$vec1[.>b]
vec2 <- 1:20
..w$vec2[.>10]
..w$iris[[nrow(.)]]
vec1 <- c(a=1,b=2,c=3,d=4)
..w$vec1[[b+c]]
vec2 <- 1:20
..w$vec2[[length(.)/2]]
}
}
