% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_grid.R
\name{err_grid}
\alias{err_grid}
\title{Make error grid for smoothed probability of location}
\usage{
err_grid(smooth = smooth, fish_data = fish_data)
}
\arguments{
\item{smooth}{smoothed daily output of forward backward filter}

\item{fish_data}{Data on individual fish, used here for multiple movement states (mvst)}
}
\value{
A list for daily error grids
}
\description{
Make error grid for smoothed probability of location
}
\examples{
h <- 1000 # 1 km resolution
set.seed(43516)
bathy <- matrix(rnorm(100, 50, 30), nrow = 10)
land <- ifelse(bathy > 0, 1, 0)
L <- array(0, dim = c(dim(bathy), 2))
L[5, 5, 1] <- 1
L[6, 6, 2] <- 1
fish_data <- data.frame("mvmt" = c(1, 1))
fwd_bck(D = 10, h = h, L = L, land = land, fish_data = fish_data)
}
