% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_1d.R
\name{est_1d}
\alias{est_1d}
\title{Estimate 1 diffusion coefficient}
\usage{
est_1d(log_D = log(D))
}
\arguments{
\item{log_D}{the log of the diffusion coefficient}
}
\value{
numeric value of the negative log likelihood
}
\description{
Provides the negative log likelihood used for estimating diffusion coefficient
}
\examples{
est_1d(log_D = log(70))
}
