% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_1d_2.R
\name{est_1d_2}
\alias{est_1d_2}
\title{Estimate 1 diffusion coefficient}
\usage{
est_1d_2(D = D, h = h, L = L, land = land)
}
\arguments{
\item{D}{the diffusion coefficient}

\item{h}{Resolution, or the length of a side in meters}

\item{L}{Likelihood for your fish over space and indexed by time}

\item{land}{Raster of land so that any probability on land can be set to zero each step}
}
\value{
numeric value of the negative log likelihood
}
\description{
Provides the negative log likelihood used for estimating diffusion coefficient
}
\examples{
est_1d(log_D = log(70), h = h, L = L, land = land)
}
