% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_2d.R
\name{est_2d}
\alias{est_2d}
\title{Estimate 1 diffusion coefficient}
\usage{
est_2d(
  log_D1 = log(D1),
  log_D2 = log(D2),
  h = h,
  L = L,
  land = land,
  fish_data = fish_data
)
}
\arguments{
\item{log_D1}{the log of the diffusion coefficient for movement state 1}

\item{log_D2}{the log of the diffusion coefficient for movement state 2}

\item{h}{Resolution, or the length of a side in meters}

\item{L}{Likelihood for your fish over space and indexed by time}

\item{land}{Raster of land so that any probability on land can be set to zero each step}

\item{fish_data}{Data on individual fish, used here for multiple movement states (mvst)}
}
\value{
numeric value of the negative log likelihood
}
\description{
Provides the negative log likelihood used for estimating diffusion coefficient
}
\examples{
est_1d(log_D = log(70), h = h, L = L, land = land)
}
