% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.r
\name{find_closest_word}
\alias{find_closest_word}
\title{Find Closest Word}
\usage{
find_closest_word(input, words)
}
\arguments{
\item{input}{A string.}

\item{words}{A vector of strings (vocabulary) to compare against the input word.}
}
\description{
Find Closest Word
}

