% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwd_bck.R
\name{fwd_bck}
\alias{fwd_bck}
\title{Forward filter for 1- or 2-state movement to estimate diffusion.}
\usage{
fwd_bck(D = D, h = h, L = L, fish.data = fish.data, land = land)
}
\arguments{
\item{D}{Diffusion coefficient for 1 or 2 movement states, can be preselected or estimated}

\item{h}{Resolution, or the length of a side in meters}

\item{L}{Likelihood for your fish over space and indexed by time}

\item{fish.data}{Data on individual fish, used here for multiple movement states (mvst)}

\item{land}{Raster of land so that any probability on land can be set to zero each step}
}
\value{
A numeric negaitve-log-likelihood value.
}
\description{
Forward filter for 1- or 2-state movement to estimate diffusion.
}
\examples{
fwd_filter(50.3)
est.D <- optim(50, fwd_filter, lower = 2, upper = 300, method = "Brent")
# need to choose upper and lower boundaries
}
