% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAEA.R
\name{getAEA}
\alias{getAEA}
\title{Convert from latitude/longitude to Albers Equal Area, same as Aleutians map (bathy.ai)}
\usage{
getAEA(lon, lat)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}
}
\value{
x and y coordinates in Albers Equal Area with names 'X' and 'Y
}
\description{
Provides the negative log likelihood used for estimating diffusion coefficient
}
