% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{L}
\alias{L}
\title{Example data}
\format{
null An array with likelihoods in each cell of the grid for each day
}
\usage{
L
}
\description{
An example likelihood from a Pacific cod with associated bathymetry file
}
\keyword{datasets}
