% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll_w_sm.R
\name{nll_w_sm}
\alias{nll_w_sm}
\title{Forward filter for 1- or 2-state movement to estimate diffusion.}
\usage{
nll_w_sm(log_D = log(D), h = h, L = L, fish_data = fish_data, land = land)
}
\arguments{
\item{h}{Resolution, or the length of a side in meters}

\item{L}{Likelihood for your fish over space and indexed by time}

\item{fish_data}{Data on individual fish, used here for multiple movement states (mvst)}

\item{land}{Raster of land so that any probability on land can be set to zero each step}

\item{D}{Diffusion coefficient for 1 or 2 movement states, can be preselected or estimated}
}
\value{
A numeric negaitve-log-likelihood value.
}
\description{
Forward filter for 1- or 2-state movement to estimate diffusion.
}
\examples{
h <- 1000 # 1 km resolution
set.seed(43516)
bathy <- matrix(rnorm(100, 50, 30), nrow = 10)
land <- ifelse(bathy > 0, 1, 0)
L <- array(0, dim = c(dim(bathy), 2))
L[5, 5, 1] <- 1
L[6, 6, 2] <- 1
fish_data <- data.frame("mvmt" = c(1, 1))
fwd_bck(D = 10, h = h, L = L, land = land, fish_data = fish_data)

}
