% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_print.R
\name{print.tag}
\alias{print.tag}
\alias{print.tag_adverb}
\title{tag and tag_adverb methods for print}
\usage{
\method{print}{tag}(x, ...)

\method{print}{tag_adverb}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\emph{tag} provides \code{tag} and \code{tag_adverb} methods for \code{print}. Additionally to
the raw code, they print the class of the object and its definition.
}
