% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_lat.R
\name{rev_lat}
\alias{rev_lat}
\title{Reverse Latitudes}
\usage{
rev_lat(m)
}
\arguments{
\item{m}{a matrix of latitude values}
}
\value{
a matrix of reversed latitude values
}
\description{
Reverse latitudes for plotting matrix as an image
}
\examples{
m <- matrix(c(50, 55, 40, 45), nrow = 2)
rev_lat(m)
}
