% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.r
\name{versioning}
\alias{versioning}
\alias{get.tagversion}
\alias{check.tagversion}
\title{TAG Versioning}
\usage{
get.tagversion()

check.tagversion(version)
}
\arguments{
\item{version}{A package version.}
}
\description{
Utilities used by the TAG system to ensure proper versioning across saved
state files.
}
