% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.r
\name{wc}
\alias{wc}
\title{Count Words of a String}
\usage{
wc(string)
}
\arguments{
\item{string}{A string.}
}
\value{
The number of words.
}
\description{
See title.
}
\details{
The basic logic is to call a word something separated by one or
more "spaces" (literal space, tab, ...).  Special treatment is
given to multiline text with a dangling dash.
}

