% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_as.R
\name{as_tag}
\alias{as_tag}
\title{Convert to a tag}
\usage{
as_tag(f)
}
\arguments{
\item{f}{a function operator}
}
\description{
convert an regular function operator to a tag
}
\details{
\code{as_tag} might not work on all cases but is made available as a
convenient way to change most adverbs into tags and benefit from
their syntax.
}
