% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_1d.R
\name{est_1d}
\alias{est_1d}
\title{Estimate 1 diffusion coefficient}
\usage{
est_1d(log_D = log(D), h = h, L = L, land = land)
}
\arguments{
\item{log_D}{the log of the diffusion coefficient}

\item{h}{Resolution, or the length of a side in meters}

\item{L}{Likelihood for your fish over space and indexed by time}

\item{land}{Raster of land so that any probability on land can be set to zero each step}
}
\value{
numeric value of the negative log likelihood
}
\description{
Provides the negative log likelihood used for estimating diffusion coefficient
}
\examples{
h <- 1000 # 1 km resolution
set.seed(43516)
bathy <- matrix(rnorm(100, 50, 30), nrow = 10)
land <- ifelse(bathy > 0, 1, 0)
L <- array(0, dim = c(dim(bathy), 2))
L[5, 5, 1] <- 1
L[6, 6, 2] <- 1
est_1d(log_D = log(10), h = h, L = L, land = land)
}
