% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_is.R
\name{is_tag_adverb}
\alias{is_tag_adverb}
\alias{is_tag}
\title{Check if a function is tag or a tag_adverb}
\usage{
is_tag_adverb(f)

is_tag(f)
}
\arguments{
\item{f}{a function}
}
\description{
Checks if the class is \code{"tag"} or \code{"tag_adverb"}
}
