% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{L}
\alias{L}
\title{Likelihood example}
\format{
A large array
}
\usage{
L
}
\description{
Example of a likelihood for a Pacific cod in the Aleutian Islands.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{details for likelihood
}\if{html}{\out{</div>}}
}
\keyword{likelihood}
