% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAG-package.r
\docType{package}
\name{textgateway-package}
\alias{textgateway-package}
\title{Text Analytics Gateway}
\description{
The Text Analytics Gateway (TAG) is an interactive webapp for performing
simple analyses on unstructured text.
}
\details{
\tabular{ll}{ 
   Package: \tab textgateway \cr 
   Type: \tab Package \cr 
   License: \tab AGPL-3 \cr 
}
}
\references{
Project home page: \url{https://github.com/XSEDEScienceGateways/textgateway}
}
\author{
Drew Schmidt
}
\keyword{Package}
