% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/textgateway-package.r
\docType{package}
\name{textgateway-package}
\alias{textgateway-package}
\title{Text Gateway}
\description{
TODO
}
\details{
\tabular{ll}{
   Package: \tab textgateway \cr
   Type: \tab Package \cr
   License: \tab AGPL-3 \cr
}
}
\author{
Drew Schmidt
}
\references{
Project home page: \url{https://github.com/XSEDEScienceGateways/textgateway}
}
\keyword{Package}

